iosDoubleSpace as InputRule #109 #242
