--- src/commands/index.js
@@ -1,13 +1,0 @@
-import {CommandSet} from 'prosemirror/dist/edit'
-import edCommands from './ed-commands'
-
-let commands = CommandSet.default
-  .add(edCommands)
-  .update(
-    { 'code:toggle': {menu: null}
-    , joinUp: {menu: null}
-    , selectParentNode: {menu: null}
-    }
-  )
-
-export default commands
